Separate `RuleBuilder`
